Machine Learning/03. Unsupervised Clustering explained/requirements.txt
pandas
scikit-learn
umap-learn
ipykernel
tqdm
requests
matplotlib
seaborn
ipywidgets
jupyter
